Calendar UI Javascript
Created applySchedule for calendar.js and tuned various templates
to make the process easier. ScheduleServer has been updated for
development and testing purposes for the calendar UI.

This is not intended to be the final form of calendar.js, but is
the first iteration of the final form. The required functionality
is implemented, but should be updated when the server's behavior
is finalized.