Removed bugs and added dev testing.
I removed a bug with the algorithm that arose when there were no
available RAs for a particular day. The bug would lead to the
candidate RA index exceeding the length of the raList.

I have also added dev testing that will allow the algorithm to be
run multiple times and display the percentage that the algorithm
was placed under review.