Updated database structure
I decided to change the name of the 'schedule' table to duties and
create a new schedule database that the duties will point to.
Hopefully this will make it easier to query against the DB since
we can just get the most recent schedule and get all of the duties
associated with that.