Converted existing functionality of editSchedule page to use fullCalendar. Also put items in place to allow users to change the duty for a selected day useing the new changeRAonDuty API hook on the scheduleServer.
Work that still remains includes updating the RA list on the right with the correct point values (possibly do this
through the new changeRAonDuty hook as it does not send back much info atm). Also need to handle errors through the
front end and it is currently possible for an AHD to assign the same RA for two of the duty slots on a weekend. This
puts that RA in an irreversable loop where if they are assigned out of the day, the new RA now takes up both slots.

Also additional functionality that is desired would be to allow AHDs to ad hoc add a duty to the calendar and assign
an RA to it immediately. Might also be nice to be able to remove a duty day as well.