Altered editSched view so that it will get accurate point value on page render.
Also fixed bug in changeRAonDuty that prevented it from finding the correct
duty. This caused nothing to be changed on the backend.

Also changed both changeRAonDuty to use getRAStats and changed the query
for getRAStats so that it always returns all of the RAs in the given hall
even if they do not have any duties for the selected time period